requirements-test.txt
coverage>=4.0.2,<5
mock>=1.3.0,<4
tox>=2.2.1,<3
djangorestframework>=3.1.1,<4
python-magic>=0.4.6
pillow>=3,<4
sorl-thumbnail>=12.3,<12.4
django-model-utils==2.3.1
django-taggit==0.18.1 
factory-boy==2.8.1

